requirements.txt
flask
stripe
flask-cors
python-dotenv
